+++ src/app/components/orders/components/process-order/process-order.component.ts
@@ -1,0 +1,13 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-process-order',
+  templateUrl: './process-order.component.html',
+  styleUrls: ['./process-order.component.css']
+})
+export class ProcessOrderComponent implements OnInit {
+
+  ngOnInit(): void {
+  }
+
+}