+++ src/app/components/orders/orders.module.ts
@@ -1,12 +1,17 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { ProcessOrderComponent } from './components/process-order/process-order.component';
+import {OrderRoutingModule} from './order-routing.module';
 
 
 
 @NgModule({
-  declarations: [],
+  declarations: [
+    ProcessOrderComponent
+  ],
   imports: [
-    CommonModule
+    CommonModule,
+    OrderRoutingModule
   ]
 })
 export class OrdersModule { }