+++ src/app/core/guards/auth.guard.ts
@@ -1,0 +1,30 @@
+import { Injectable } from '@angular/core';
+import type { ActivatedRouteSnapshot, CanActivate, RouterStateSnapshot, UrlTree } from '@angular/router';
+import { Observable } from 'rxjs';
+import {AuthService} from '../services/auth.service';
+import {Router} from '@angular/router';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class AuthGuard implements CanActivate {
+  constructor(
+    private authService: AuthService,
+    private router: Router
+  ) {}
+
+  canActivate(
+    route: ActivatedRouteSnapshot,
+    state: RouterStateSnapshot): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
+    const { url } = state;
+    return this.checkLogin(url);
+  }
+
+  private checkLogin(url: string): boolean | UrlTree {
+    if (this.authService.isLoggedIn) { return true; }
+// Store the attempted URL for redirecting
+    this.authService.redirectUrl = url;
+// Navigate to the login page, return UrlTree
+    return this.router.parseUrl('/login');
+  }
+}